requirements.txt
requests>=2.31.0
Pillow>=10.0.0
pyinstaller>=6.0.0
psutil>=5.9.0
tqdm>=4.66.0
colorama>=0.4.6
watchdog>=3.0.0
python-dateutil>=2.8.0